Re #189: Implemented fuzzy decision making for automatic oversampling calculation. Sensible oversampling factors are calculated for the used test data, although the membership functions could use some adjustment.
git-svn-id: https://subversion.assembla.com/svn/slicerrt/trunk/SlicerRt/src@2164 25a7314f-c4c8-4314-9936-a25f8480a772

+++ Contours/MRML/vtkCalculateOversamplingFactor.h
@@ -79,8 +79,8 @@ public:
   vtkGetObjectMacro(RasterizationReferenceVolumeNode, vtkMRMLScalarVolumeNode);
   vtkSetObjectMacro(RasterizationReferenceVolumeNode, vtkMRMLScalarVolumeNode);
 
-  vtkGetMacro(OutputOversamplingFactor, int);
-  vtkSetMacro(OutputOversamplingFactor, int);
+  vtkGetMacro(OutputOversamplingFactor, double);
+  vtkSetMacro(OutputOversamplingFactor, double);
 
   vtkGetMacro(LogSpeedMeasurements, bool);
   vtkSetMacro(LogSpeedMeasurements, bool);
@@ -104,7 +104,7 @@ protected:
   double ComplexityScalingFactor;
 
   /// Calculated oversampling factor for the contour node and its reference volume
-  int OutputOversamplingFactor;
+  double OutputOversamplingFactor;
 
   /// Flag telling whether the speed measurements are logged on standard output
   bool LogSpeedMeasurements;
